+++ common/cba2.c
@@ -90,7 +90,7 @@ void generateCMake(BuildInfo* buildInfo) {
     }
     
     // Add subdirectories automatically
-    char* directories = ListDirectories(".");
+    char* directories = getListedDirectories(".");
     if(directories) {
         char* dir = strtok(directories, "\n");
         while(dir) {