+++ common/cxtMap.h
@@ -2,6 +2,7 @@
 #define CXT_MAP_H
     #include <stdbool.h>
     #include <stddef.h>
+    #include <string.h>
 
     typedef struct {
         char* key;