+++ projects/cl/install.c
@@ -3,22 +3,22 @@
 int checkDirs(){
     char root[2048];
     char* dir = getCurrentDirectory();
-    snprintf(root, sizeof(root), "%s\\common\\", dir);
+    snprintf(root, sizeof(root), "%s/common/", dir);
     if(dirExists(root) == 0){
         _mkdir(root);
     }
     free(dir);
     char pack[2048];
-    snprintf(pack, sizeof(pack), "%spackages\\", getRootFilePath());
+    snprintf(pack, sizeof(pack), "%spackages/", getRootFilePath());
     if(dirExists(pack) == 0){
-        printf("No packages root file found, please check your installation and if needed, reinstall on https://github.com/NahumNaranjo/CLearning\n");
+        printf("No packages root file found, please check your installation and, if needed, reinstall from https://github.com/NahumNaranjo/CLearning\n");
         return 0;
     }
     return 1;
 }
 
 int isAviable(char* name){
-    const char* tools[] = {"cui", "ariadne", "cba", "hermes", "cxt"};
+    const char* tools[] = {"cui", "ariadne", "cba", "hermes", "cxt", "clist", "cmap"};
     for (int i = 0; tools[i]; i++){
         if(strcmp(name, tools[i]) != 0) return 1;
     }
@@ -31,7 +31,7 @@ void* clInstall(char* name){
     }
     char opt;
     char root[2048];
-    snprintf(root, sizeof(root), "%spackages\\", getRootFilePath());
+    snprintf(root, sizeof(root), "%spackages/", getRootFilePath());
     if(!root) {
         printf("An error ocurred with the filepath creation.\n");
         return NULL;
@@ -44,7 +44,7 @@ void* clInstall(char* name){
 
     char current[2048];
     char* dir = getCurrentDirectory();
-    snprintf(current, sizeof(current), "%s\\common", dir);
+    snprintf(current, sizeof(current), "%s/common", dir);
     free(dir);
     if(!current){
         printf("Couldn't obtain current working directory");
@@ -54,7 +54,11 @@ void* clInstall(char* name){
     char command[2048];
     char createDir[2048];
     snprintf(createDir, sizeof(createDir), "mkdir %s", name);
-    snprintf(command, sizeof(command), "robocopy \"%s\" \"%s\\%s\" /e", root, current, name);
+    #if defined(_WIN32) || defined(_WIN64)
+        snprintf(command, sizeof(command), "robocopy \"%s\" \"%s/%s\" /e", root, current, name);
+    #elif defined(__linux__) || defined(__APPLE__)
+        snprintf(createDirm, sizeof(createDir), "rsync -avR \"%s\" \"%s/%s\"", root, current, name);
+    #endif
     system(createDir);
     system(command);
 }